Remove internal/kptops dead code
