--- internal/kptops/functions.go
@@ -1,29 +1,0 @@
-// Copyright 2022 The kpt Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package kptops
-
-import (
-	"sigs.k8s.io/kustomize/kyaml/fn/framework"
-)
-
-var functions map[string]framework.ResourceListProcessorFunc = map[string]framework.ResourceListProcessorFunc{
-	"ghcr.io/kptdev/krm-functions-catalog/apply-setters:v0.2.0": applySetters,
-	"ghcr.io/kptdev/krm-functions-catalog/set-labels:v0.1.5":    setLabels,
-	"ghcr.io/kptdev/krm-functions-catalog/set-namespace:v0.4.1": setNamespace,
-}
-
-func FindProcessor(image string) framework.ResourceListProcessorFunc {
-	return functions[image]
-}
