--- internal/kptops/testdata/render-with-function-config/simple-bucket/Kptfile
@@ -1,26 +1,0 @@
-# Copyright 2022 The kpt Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: kpt.dev/v1
-kind: Kptfile
-metadata:
-  name: simple-bucket
-  annotations:
-    blueprints.cloud.google.com/title: Google Cloud Storage Bucket blueprint
-info:
-  description: A Google Cloud Storage bucket
-pipeline:
-  mutators:
-    - image: ghcr.io/kptdev/krm-functions-catalog/apply-setters:v0.2.0
-      configPath: setters.yaml
