--- internal/kptops/testdata/render-with-inline-config/expected.txt
@@ -1,56 +1,0 @@
-apiVersion: config.kubernetes.io/v1
-kind: ResourceList
-items:
-- apiVersion: kpt.dev/v1
-  kind: Kptfile
-  metadata:
-    name: simple-bucket
-    annotations:
-      blueprints.cloud.google.com/title: Google Cloud Storage Bucket blueprint
-      config.kubernetes.io/index: '0'
-      config.kubernetes.io/path: 'Kptfile'
-      internal.config.kubernetes.io/index: '0'
-      internal.config.kubernetes.io/path: 'Kptfile'
-      internal.config.kubernetes.io/seqindent: 'wide'
-  info:
-    description: A Google Cloud Storage bucket
-  pipeline:
-    mutators:
-    - image: ghcr.io/kptdev/krm-functions-catalog/apply-setters:v0.2.0
-      configMap:
-        name: updated-bucket-name
-        namespace: updated-namespace
-        project-id: updated-project-id
-        storage-class: updated-storage-class
-# Copyright 2022 The kpt Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-- apiVersion: storage.cnrm.cloud.google.com/v1beta1
-  kind: StorageBucket
-  metadata: # kpt-merge: config-control/blueprints-project-bucket
-    name: updated-project-id-updated-bucket-name # kpt-set: ${project-id}-${name}
-    namespace: updated-namespace # kpt-set: ${namespace}
-    annotations:
-      cnrm.cloud.google.com/force-destroy: "false"
-      cnrm.cloud.google.com/project-id: updated-project-id # kpt-set: ${project-id}
-      cnrm.cloud.google.com/blueprint: 'kpt-fn,krm-fn'
-      config.kubernetes.io/index: '0'
-      config.kubernetes.io/path: 'bucket.yaml'
-      internal.config.kubernetes.io/index: '0'
-      internal.config.kubernetes.io/path: 'bucket.yaml'
-      internal.config.kubernetes.io/seqindent: 'compact'
-  spec:
-    storageClass: updated-storage-class # kpt-set: ${storage-class}
-    uniformBucketLevelAccess: true
-    versioning:
-      enabled: false
