--- internal/kptops/testdata/render-with-inline-config/simple-bucket/Kptfile
@@ -1,16 +1,0 @@
-apiVersion: kpt.dev/v1
-kind: Kptfile
-metadata:
-  name: simple-bucket
-  annotations:
-    blueprints.cloud.google.com/title: Google Cloud Storage Bucket blueprint
-info:
-  description: A Google Cloud Storage bucket
-pipeline:
-  mutators:
-    - image: ghcr.io/kptdev/krm-functions-catalog/apply-setters:v0.2.0
-      configMap:
-        name: updated-bucket-name
-        namespace: updated-namespace
-        project-id: updated-project-id
-        storage-class: updated-storage-class
