--- internal/kptops/testdata/render-with-inline-config/simple-bucket/bucket.yaml
@@ -1,28 +1,0 @@
-# Copyright 2022 The kpt Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: storage.cnrm.cloud.google.com/v1beta1
-kind: StorageBucket
-metadata: # kpt-merge: config-control/blueprints-project-bucket
-  name: updated-project-id-updated-bucket-name # kpt-set: ${project-id}-${name}
-  namespace: updated-namespace # kpt-set: ${namespace}
-  annotations:
-    cnrm.cloud.google.com/force-destroy: "false"
-    cnrm.cloud.google.com/project-id: updated-project-id # kpt-set: ${project-id}
-    cnrm.cloud.google.com/blueprint: 'kpt-fn'
-spec:
-  storageClass: updated-storage-class # kpt-set: ${storage-class}
-  uniformBucketLevelAccess: true
-  versioning:
-    enabled: false
